+++ Sources/Sentry/SentryUIViewControllerPerformanceTracker.m
@@ -5,7 +5,6 @@
 #import "SentrySDK+Private.h"
 #import "SentryScope.h"
 #import "SentrySpanId.h"
-#import "SentrySwizzle.h"
 #import "SentryUIViewControllerSanitizer.h"
 #import <objc/runtime.h>